pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ftune"
version = "0.1.0"
description = "Estimate GPU memory, training time, and costs for LLM fine-tuning"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
keywords = ["llm", "fine-tuning", "gpu", "memory", "cost-estimator", "lora", "qlora"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
]
cli = [
    "typer>=0.12.0",
    "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ftune"
Documentation = "https://github.com/yourusername/ftune#readme"
Repository = "https://github.com/yourusername/ftune"
Issues = "https://github.com/yourusername/ftune/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ftune"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 100
